requirements.txt
aws-cdk-lib==2.8.0